requirements.txt
torch_geometric==2.3.0
tensorboard==2.13.0
atom3d==0.2.6
jupyter==1.0.0 
notebook==6.5.4
ipywidgets==8.0.6
lightning==2.0.2
e3nn==0.5.1
matplotlib==3.7.1
plotly==5.14.1
# Adjust the wheel with the correct CUDA version and OS:
# https://data.pyg.org/whl/torch-2.0.0+cu117/torch_cluster-1.6.1+pt20{CUDA-VERSION}-{PY-VERSION}-{OS}.whl
torch_cluster 
https://data.pyg.org/whl/torch-2.0.0+cu117/torch_cluster-1.6.1+pt20cu117-cp311-cp311-linux_x86_64.whl
torch_scatter 
https://data.pyg.org/whl/torch-2.0.0+cu117/torch_scatter-2.1.1+pt20cu117-cp311-cp311-linux_x86_64.whl
torch_sparse 
https://data.pyg.org/whl/torch-2.0.0+cu117/torch_sparse-0.6.17+pt20cu117-cp311-cp311-linux_x86_64.whl
